requirements.txt
yfinance
pandas
streamlit
pymongo
python-dotenv
matplotlib
